# perf: support strict union