--- a/src/validators/union.rs
+++ b/src/validators/union.rs
@@ -1,7 +1,7 @@
 use pyo3::prelude::*;
 use pyo3::types::{PyDict, PyList};
 
-use crate::build_tools::SchemaDict;
+use crate::build_tools::{is_strict, SchemaDict};
 use crate::errors::{LocItem, ValError, ValLineError};
 use crate::input::Input;
 
@@ -10,6 +10,7 @@ use super::{build_validator, BuildContext, BuildValidator, CombinedValidator, Ex
 #[derive(Debug, Clone)]
 pub struct UnionValidator {
     choices: Vec<CombinedValidator>,
+    strict: bool,
 }
 
 impl BuildValidator for UnionValidator {
@@ -25,7 +26,8 @@ impl BuildValidator for UnionValidator {
             .iter()
             .map(|choice| build_validator(choice, config, build_context).map(|result| result.0))
             .collect::<PyResult<Vec<CombinedValidator>>>()?;
-        Ok(Self { choices }.into())
+        let strict = is_strict(schema, config)?;
+        Ok(Self { choices, strict }.into())
     }
 }
 
@@ -37,33 +39,52 @@ impl Validator for UnionValidator {
         extra: &Extra,
         slots: &'data [CombinedValidator],
     ) -> ValResult<'data, PyObject> {
-        // 1st pass: check if the value is an exact instance of one of the Union types
-        if let Some(res) = self
-            .choices
-            .iter()
-            .map(|validator| validator.validate_strict(py, input, extra, slots))
-            .find(ValResult::is_ok)
-        {
-            return res;
+        if self.strict {
+            let mut errors: Vec<ValLineError> = Vec::with_capacity(self.choices.len());
+
+            for validator in &self.choices {
+                let line_errors = match validator.validate_strict(py, input, extra, slots) {
+                    Err(ValError::LineErrors(line_errors)) => line_errors,
+                    otherwise => return otherwise,
+                };
+
+                errors.extend(
+                    line_errors
+                        .into_iter()
+                        .map(|err| err.with_prefix_location(LocItem::S(validator.get_name(py)))),
+                );
+            }
+
+            Err(ValError::LineErrors(errors))
+        } else {
+            // 1st pass: check if the value is an exact instance of one of the Union types
+            if let Some(res) = self
+                .choices
+                .iter()
+                .map(|validator| validator.validate_strict(py, input, extra, slots))
+                .find(ValResult::is_ok)
+            {
+                return res;
+            }
+
+            let mut errors: Vec<ValLineError> = Vec::with_capacity(self.choices.len());
+
+            // 2nd pass: check if the value can be coerced into one of the Union types
+            for validator in &self.choices {
+                let line_errors = match validator.validate(py, input, extra, slots) {
+                    Err(ValError::LineErrors(line_errors)) => line_errors,
+                    otherwise => return otherwise,
+                };
+
+                errors.extend(
+                    line_errors
+                        .into_iter()
+                        .map(|err| err.with_prefix_location(LocItem::S(validator.get_name(py)))),
+                );
+            }
+
+            Err(ValError::LineErrors(errors))
         }
-
-        let mut errors: Vec<ValLineError> = Vec::with_capacity(self.choices.len());
-
-        // 3rd pass: check if the value can be coerced into one of the Union types
-        for validator in &self.choices {
-            let line_errors = match validator.validate(py, input, extra, slots) {
-                Err(ValError::LineErrors(line_errors)) => line_errors,
-                otherwise => return otherwise,
-            };
-
-            errors.extend(
-                line_errors
-                    .into_iter()
-                    .map(|err| err.with_prefix_location(LocItem::S(validator.get_name(py)))),
-            );
-        }
-
-        Err(ValError::LineErrors(errors))
     }
 
     fn get_name(&self, _py: Python) -> String {