--- a/tests/validators/test_union.py
+++ b/tests/validators/test_union.py
@@ -215,3 +215,17 @@ def test_no_choices():
         SchemaValidator({'type': 'union'})
 
     assert exc_info.value.args[0] == ('Error building "union" validator:\n' '  KeyError: \'"choices" is required\'')
+
+
+def test_strict_union():
+    v = SchemaValidator({'type': 'union', 'strict': True, 'choices': [{'type': 'bool'}, {'type': 'int'}]})
+    assert v.validate_python(1) == 1
+    assert v.validate_python(123) == 123
+
+    with pytest.raises(ValidationError) as exc_info:
+        v.validate_python('123') == 123
+
+    assert exc_info.value.errors() == [
+        {'kind': 'bool_type', 'loc': ['bool'], 'message': 'Value must be a valid boolean', 'input_value': '123'},
+        {'kind': 'int_type', 'loc': ['int'], 'message': 'Value must be a valid integer', 'input_value': '123'},
+    ]