```
main branch
———
[1_000_000 times] lax union with 5 choices: 2.89s
[1_000_000 times] strict union with 5 choices: 2.84s

strict-union branch
——————
[1_000_000 times] lax union with 5 choices: 2.84s
[1_000_000 times] strict union with 5 choices: 1.84s
```

<details>
  <summary>Run script</summary>
  
```py
import time
from pydantic_core import SchemaValidator, ValidationError


for strict in (False, True):
    v = SchemaValidator({
        'type': 'model',
        'config': {'strict': strict},
        'fields': {
            'x': {
                'schema': {
                    'type': 'union',
                    'choices': [{'type': 'bool'}, {'type': 'int'}, {'type': 'str'}, {'type': 'float'}, {'type': 'date'}],
                }
            }
        }
    })
    start = time.time()
    for _ in range(1_000_000):
        try:
            v.validate_python({'x': [0]})
        except ValidationError as e:
            pass
        else:
            raise
    print(f"[1_000_000 times] {'strict' if strict else 'lax'} union with 5 choices: {time.time() - start:.2f}s")
```
</details>