@PrettyWood 2022-06-21
I can add a benchmark if you want sure! And I felt like having a dedicated validator for this was overkill but I may be wrong. I can probably check tonight