@samuelcolvin 2022-06-21
> And I felt like having a dedicated validator for this was overkill but I may be wrong. I can probably check tonight

I think you're probably right, I'm just intrigued about what the performance change is. If it's just cutting out one bool check it'll be virtually zero, but sometimes small changes like this allow the compiler to do significantly more optimisations and you see a big improvement.