@PrettyWood 2022-06-21
I'll check tonight! Small remark: wouldn't it make more sense to have `strict` in the field information (outside the schema with `default`, `required`, ...)?