@PrettyWood 2022-06-22
Run with first implem
```
------------------------------------------ benchmark 'smart-union': 2 tests -----------------------------------------
Name (time in ns)                Min                  Mean             StdDev            Outliers  Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------
test_smart_union_core       127.0795 (1.0)        128.9181 (1.0)       3.2905 (1.0)      406;4130   91258         100
test_smart_union_pyd      1,595.7979 (12.56)    1,636.6685 (12.70)    30.9371 (9.40)     824;1744   62500          10
---------------------------------------------------------------------------------------------------------------------

-------------------------------------------- benchmark 'strict-union': 2 tests --------------------------------------------
Name (time in ns)                 Min                  Mean                StdDev              Outliers  Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------
test_strict_union_core       749.9999 (1.0)        908.5145 (1.0)        542.6710 (1.0)      8417;19309  147233          10
test_strict_union_pyd      3,770.4885 (5.03)     4,085.4826 (4.50)     7,437.9480 (13.71)      251;3905  130430           2
---------------------------------------------------------------------------------------------------------------------------
```

Run with second implem
```
------------------------------------------ benchmark 'smart-union': 2 tests -----------------------------------------
Name (time in ns)                Min                  Mean             StdDev            Outliers  Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------
test_smart_union_core       148.3300 (1.0)        152.0422 (1.0)       3.6696 (1.0)      687;9828   93757         100
test_smart_union_pyd      1,587.4975 (10.70)    1,631.3385 (10.73)    82.2311 (22.41)    342;1353   62665          10
---------------------------------------------------------------------------------------------------------------------

------------------------------------------ benchmark 'strict-union': 2 tests -------------------------------------------
Name (time in ns)                 Min                  Mean              StdDev             Outliers  Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------
test_strict_union_core       691.6001 (1.0)        724.1021 (1.0)       22.0824 (1.0)      1748;1610  149077          10
test_strict_union_pyd      3,833.0036 (5.54)     4,022.7391 (5.56)     110.7831 (5.02)     2634;6575  128354           2
------------------------------------------------------------------------------------------------------------------------
```

@samuelcolvin Should be good!