@PrettyWood 2022-06-22
@samuelcolvin I root for the first implementation (HEAD^) since the perf gain is only on strict unions with errors. I think we expect more successful union validations than errors